+++ pytsite/util/__init__.py
@@ -1,8 +1,8 @@
 """PytSite Helpers.
 """
+# Public API
+from ._functions import *
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
-# Public API
-from ._functions import *